requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-anthropic>=0.1.15
anthropic>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
PyGithub>=2.1.0
gitpython>=3.1.0
pyyaml>=6.0.0
jinja2>=3.1.0
typing-extensions>=4.9.0
